PostgreSQL Convergence Logic
Convergence is the process of converging from an arbitrary physical state of
PostgresResource to its target state, which is combination of target vm size,
target storage size and target server count.

Convergence logic is implemented in 4 steps:
1. Provision enough `fresh` servers to meet the target server count.
2. Wait until new servers are `ready` (i.e. in wait state).
3. Recycle representative_server to match resource's target.
4. Prune extra servers that are possibly left over from pre-convergence time.

At the end of the convergence, we ensure that the representative_server is in
the desired state and there are enough fresh standby servers that are ready to
take over the primary role if it becomes necessary.

Since the convergence is long running operation, it is possible that the target
changes during the convergence, especially during second step as other steps
are relatively quick. To ensure converging to correct target, second step can
decide to hop back to first step for course correction.