Use rhizome user for service VM sshables
Use rhizome as unix_user for sshables associated with VMs, with the
exception of the runneradmin user.

When bootstrapping rhizome, add operator_ssh_public_keys to authorized
keys of the bootstrap user.

For VM hosts, the operator/bootstrap user is `root`. For service VMs,
the user used for bootstrapping and operator actions is configurable. By
convention, that user is `ubi`.