Validate VAT async when it's changed
VAT validation is a costly operation, and also the European Commission
services are fragile. So, it's better to validate VAT in the background
without blocking the customer. If it fails, we can manually send an
email to the customer to inform them about the failure for now. Once we
trust the operation, we can automate the email notifications.