Improve display of option values in CLI
Show available option values in separate section after options display.
In cases where Options can be taken from elsewhere in the application,
do that instead of hard coding the options.  This allows burstable-1
and burstable-2 to show up as VM/PG size options, and show more
available VM/PG storage sizes. This doesn't show which storage sizes
are valid for a given VM/PG size.  That's possible but would clutter
the display.

Update to a later rodish checkout, which moved some of the features
UbiCli uses to plugins, and load those plugins.