Do not include CLI --confirm option in help output
Now that it is the only option left, we can suppress the display
of the entire options section for the root command.  I think this
is fine to hide as users are not supposed to use it directly, it
is used implicitly in destroy command handling to receive user
confirmation.

Update usage for root command so it does not show options, and it
shows that a command is required instead of optional.