Avoid potentially deleting a cache entry just updated during a race condition
With PostgreSQL's default READ COMMITTED transaction isolation level,
I believe there is a race condition previously.  Dataset#destroy here
does a SELECT to get the rows, and then individual DELETE statements
for each.  If the rows are not locked by the SELECT query, then they
can be updated concurrently before the DELETE.  The most common UPDATE
would be an update in the last_accessed_at field to set it to
CURRENT_TIMESTAMP, in which case the cache entry should not be deleted.

Using FOR UPDATE here will lock the rows, so they cannot be
concurrently updated.  This ensures if the rows are retrieved by the
SELECT, they are safe to DELETE.  A concurrent UPDATE would block,
and after the DELETE finished, the UPDATE would no longer include
the row.

While here, switch from Dataset#each to Dataset#all in the loop,
because Sequel does not support queries inside Dataset#each (though
it does work with the postgres adapter when not streaming).

Also, change the break condition to be at the end of the loop instead
of the beginning.  We already check the same condition before the
loop so this avoids one unnecessary check and loop iteration.