Extend the runner's destroy deadline if it's still running a job
Before destroying the runner, we also deregister it from GitHub to
ensure it's not running any jobs.

Sometimes, GitHub assigns a job to the runner at the last minute when we
decide to recycle it.

There's nothing we can do about it, so we just wait for the job to
finish.

We simply extend the existing 15-minute deadline and set a new deadline
for 2 hours.