Add VPC related resources table and flex null restrictions - Migration
Here we are adding a new table called private_subnet_aws_resource to
keep track of the resource ids on the AWS side that simply makes up a
functioning private subnet on the Ubicloud side. That is especially
important to clean-up after the destroy operation.

We are also making it a bit more flexible to add resources to the system
by allowing assigned_vm_address.address_id to be nil since it doesn't
map to a pre provisioned subnet and vm_storage_volume related references
since these are not referencing a host now. They are basically resources
that comes with a VM and goes when the VM is destroyed.