Provision AWS VPC instead of Ubicloud Private Subnets
Here we are adding the first forking logic into the subnet_nexus.rb. To
be able to provision a private subnet like thing in AWS, we need the
following resources:
- VPC
- Subnet in that VPC
- Security Group to be able to manage firewalls
- InternetGateway to be able to access public network
- RouteTable to be able to manage the traffic in the vpc.

All of these resources are created in a controlled manner in a different
file so that the main logic in the nexus.rb files are left separate.