Provision AWS Network Interface instead of Ubicloud NICs
Nics are logical entities in Ubicloud which we use to configure TAP
devices later on. In AWS, they are similar, but they also get the public
IP addresses assigned. In Ubicloud, they are mostly related to private
networking. Here in this commit, we introduce the logic to prepare an
AWS network interface to be used by a VM. The only missing piece is the
private IPv6 element. We'll skip that for now.