Integrate PG with AWS instance sizes
On the AWS side, we are limited on the VM size configuration with disk
sizes. Therefore, we are introducing the limitations here. We are also
making only standard SKU available. For now, we will only support a
single disk size per instance size which also is supported up to 16
vcpu.