Check that provided locations in routes are valid visible locations
Add LocationNameConverter.to_visible_internal_name to check that a
given location is a valid visible location.

Add Clover#check_visible_location, which calls to_visible_internal_name
and halts route processing if a location is not a valid visible location.
Use this in the routes in all cases where @location was being set
manually.

Change delete specs that use invalid locations to expect 404 instead of
204.  204 is not needed, as the deletion could never have succeeded with
an invalid location.