Make the GitHub resource cleanup more reliable in E2E
Since the GitHub repository has a nexus, we should use the destroy
semaphore instead of deleting it directly.

We destroy cache entries when we destroy the repository, so it's better
to let nexus handle it.

We also need to ensure that all VM pools and repositories are deleted
before removing the related projects.