Record runner allocation timestamp
Right now, it's difficult to see how long each runner waited for
allocation. If a customer hits the concurrency limit, they wait longer.

This timestamp will help us understand the historical impact of
concurrency limits on customer experience.

The VM has a similar column for when it has a VM host, and the runner
has a column for when it has a VM.