Log the time spent waiting for the runner concurrency limit
It prints whether allocation is allowed or not due to the concurrency
limit in every iteration, but it's hard to tell how long it waited.

It logs duration with `runner_capacity_waited` tag now.

I also combined two `hop_allocate_vm` calls in `wait_concurrency_limit`
into one to make the logging a single line. There are no changes in
behavior.