Nap 6 hours while waiting for the semaphore to increase
When we increase the semaphore, we have already scheduled the strand for
now.

If the label is just waiting for the semaphore to increase, there's no
need for short naps.

Most of our wait labels are just waiting for the semaphore to increase,
so I extended their naps to 6 hours.

It will help decrease the load on the respirate on production

[^1]: https://github.com/ubicloud/ubicloud/blob/28dacb968b9222e72bdc829c240ad3d272fb80c8/model/semaphore.rb#L10