Adjust Billing Rates for k8s
We will start charging for k8s clusters, so this commit updates
the billing rates for the nodes.

We don't offer node selection choices at the moment and default to
the smallest node size standard-2 with public IP and 40GB storage.
The pricing calculation for a k8s node is the same with a bare VM
with the same specs.

We might need to adjust slightly when we offer different sizes for
nodes. Since we don't have a separate line item for the Public IP
of a node, the pricing of k8s nodes won't scale the same with VMs.