Add auto_validations skip_invalid option
`skip_invalid` skips auto validations on a column if the column is
already invalid for other reasons.  This can prevent the display
of multiple error message for a column, where only one is needed.
This affects one spec, which had both a detailed custom error message,
as well as a generic "is not present" error message.

I tried using `not_null: :presence`, which validates that NOT NULL
string columns do not have empty string values, but Ubicloud has a
lot of cases where NOT NULL string columns are allowed to have empty
string values, so it doesn't make sense to use that.