Fix loadbalancer tests flakiness
After introducing the multiple ports feature, some tests started to
fail randomly

The issue was calling some associations like vm_ports on the load_balancer
model for example. we should not rely on the first and last since the result
is unordered. Since there was no order the generated rules would differ if
compared char-to-char but had the same functionality.

LB methods are modified to generate rules based on the sorted results
(sorted whether on port or address) which gives deterministic output

Also the test specs are improved to use the database records instead of
mocking for a cleaner code. for example Nics are now created instead of
being mocked at every level.