Disallow access to pods GUA ips from internet
So far pods had both ULA and GUA ipv6 ips and the GUA ip was accessible from
the internet. with this change, we add nft rules to each vm to only allow
connections to be made from inside the pod and disallow all new connections
coming to the GUA ip address.

This change is critical to not expose pods to the internet.