Try to charge the latest payment method
The payment method has an `order` column. I intended to add this feature
initially to allow customers to prioritize their payment methods, so
they can select a primary and a backup option. However, we haven't added
it to the UI yet, so all `order` columns are currently nil.

When a payment fails, customers simply log in to the console and add a
new payment method while keeping the old one. When we try to charge the
invoice, we start with the oldest method, which probably isn't working
anymore. We should charge from the newest method to the oldest to reduce
our payment failure rate.