Parse DockerHub rate limit output
We started logging the remaining DockerHub rate limit output to track
our usage since DockerHub plans to change its rate limit policy.

Right now, it prints all output as raw text, making it difficult to
analyze in our monitoring tool.

It would be better to log it as JSON so we can run basic queries on it.

Example output to parse:

    ratelimit-limit: 100;w=21600
    ratelimit-remaining: 100;w=21600
    docker-ratelimit-source: 192.168.1.1