Bump maximum pagination page size to 1000, and make that the default
Actually testing over 1000 results in the specs is way too slow,
so just test that a page_size lower than the number of objects
is respected.  This doesn't test the upper clamp, but it's not
necessary to do so for coverage, and I don't think the benefit
of such testing is worth the large performance hit.