Avoid Roda::RodaPlugins::Sessions::CookieTooLarge for unauthenticated request to long path
Clover uses Rodauth's login_return_to_requested_location? feature,
which store the path to redirect to after login in the session.
Unfortunately, for too long paths, this results in exceeding the
4K cookie size limit, which is an unhandled production exception.

I fixed this in Rodauth's master branch, by enforcing a limit for
the path size, so update to the current Rodauth master branch to
fix this issue.