Handle case when user attempts to add a user to a project when they already have access
This raises Sequel::UniqueConstraintViolation, so rescue that and
show an appropriate error message.  As the specs use transactions,
you need to use a savepoint in the specs to avoid issues.  This
uses the savepoint: :only option, so it will open a savepoint
during the specs, but not create a new transaction outside of the
specs.