Deduplicate service ports by 'port' value, keeping first occurrence
Previously customers were able to create multiple LoadBalancer services
with duplicate ports which would cause issue for us.

The proper way to do it is installing a webhook inside the clustere and
not allow the creation of the duplicate ports.

As a quick fix for now, we would sort all LoadBalancer services based on
the creationTimestamp and use the first occurance and ignore the other ones.