Do not calculate VAT if the total is below the minimum charge threshold
If the total is below the minimum charge threshold, we don’t charge the
customer. We won’t collect VAT for it, so there’s no need to calculate
or show it on the invoice.

I moved the VAT assignment to after the cost calculation since we need
the total cost to decide whether to calculate it or not.