Set a default Strand lease of now() - 1000 years
Also, backfill the column. It sort of "naturally" backfills as leases
are acquired and cleared using the new code in
a05fc1db6159708240ef83b43f19f2e79b5decf7, but by using a default and a
one-time backfill after that the code can move towards a `NOT NULL`
constraint.