Add support for respirate partitioning
This adds a partition keyword argument to the dispatcher, which accepts
a range of UUID strings.  If provided, the dispatcher runs in
partitioning mode.  In partitioning mode:

* The default scan query returns only strands inside the partition.

* There is a scan_old method that can be used to return strands where
  schedule is more than 5 seconds in the past.  Respirate runs this
  scan every 5 seconds.  This handles failure of one paritioned
  respirate process gracefully.  Other respirate processes will pick
  up the slack, just with an additional 5 seconds of latency.

The advantage of partitioning is that multiple respirate processes
no longer try to operate on the same strands, so this reduces
contention.

For bin/respirate usage, provide 2 arguments to enable partitioning.
The first argument is the number of partitions, and the second is the
partition number for each partition (1 based). So:

  partition 4 2

Divides the UUID space into 4 partitions, and operates on the 2nd of 4
("80000000-0000-0000-0000-000000000000"..."c0000000-0000-0000-0000-000000000000").

For Heroku usage, the second argument is optional if the DYNO environment
variable is present, in which case it will try to parse the DYNO value
to determine the current partition.

Currently, up to 16 partitions are supported, and it is trivial to go
from 16 to 256 if needed.

This should be compatible with an index-only scan if we have an index
on `strand(schedule, lease, id) WHERE exitval IS NULL`, once the
`lease IS NULL` condition is removed.  The scan_old query should also
be able to use the same index for an index-only scan.

While here, add a minor spec fix to shutdown the dispatcher when using
custom dispatcher options.