Fix LoadBalancer cert renewal empty cert bug
Recently we have hit the issue of certificate being empty in inference
endpoints. This happened after a cert renewal. The bug is that, the
moment we create a new CertNexus, created_at column is set by default
but the cert itself is not finalized just yet. Since we forgot to check
that column, we thought the cert is finalized and went into broadcasting
the new certificate to the metadata-endpoint. Now, checking if the cert
column is properly filled, we essentially wait for the new certificate
to be finalized before going into distribution mode.