Shorten GitHub app JWT lifetime
While generating the JWT token for our GitHub app client, we set the
expiration to 10 minutes, which is the maximum allowed lifetime for a
JWT token [^1]

Since our servers and GitHub's server clocks aren't perfectly
synchronized, we occasionally see the following exception in the
production environment:

    Octokit::Unauthorized - POST https://api.github.com/app/installations/***/access_tokens: 401 - 'Expiration time' claim ('exp') is too far in the future // See: https://docs.github.com/rest

Our lease time is 120 seconds, so it’s fine to shorten the JWT lifetime
to any value over 120 seconds. To be safe for both sides, I set it to 8
minutes.

[^1]: https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/generating-a-json-web-token-jwt-for-a-github-app