Do not show runners that are already being destroyed
Sometimes, runner virtual machines get stuck in the destroy label due to
host crashes or similar issues. These stuck runners confuse our
customers, so let's not display them in the UI.

A better solution would be to disassociate the runner and VM when the
destroy process starts. However, this requires more work since the VM
pool picks virtual machines that don't have runners. I will work on this
in the future.