Do not wait 5 seconds when processing old strands
If we picked up old strands during the last old strand scan, that
means there could be a backlog.  Do not wait 5 seconds in this case
to scan for more old strands.  Instead, continuing checking for
both the current partition strands and old strands in every loop
until there are no old strands.  Additionally, do not sleep for 1
second unless there are no strands for the current partition and no
old strands.  This results in additional checks for strands for the
current partition, even if the there were no strands for the current
partition in the last scan, if there are old strands to process,
but I think that's acceptable.