Switch firewall creation form to a simpler approach using forme
This uses the forme library to build the form. Forme is an HTML form
library that has integrations for both Roda and Sequel (it's older than
Roda, first release was in 2011).  In addition to simplifying HTML form
creation, another benefit is it handles CSRF tags implicitly.

For the javascript showing and hiding of inputs based on selected options,
this adds a simpler 10 line implementation to replace the ~60 line current
implementation.  The new implementation does not require any inline
javascript to set option_tree, option_children, and option_dirty.
Instead:

* The form lists what types of dependencies it supports
  (`dependencies: "location"`)

* The radio button set to choose the parent object marks itself as a
  provider (`provides: "location"`).

* Any elements that have a dependency on a parent have 2 classes, one
  of which is the value of the radio button for the related dependency,
  and the other is `depends-` followed by the dependency name
  (`class: "depends-location #{_1.location_id}"`).

On a change to a radio button, all elements with the `depends-*` class
are hidden, and then the ones with a class matching the value of the
radio button are shown.

This approach also does not require the use of the OptionTreeGenerator or
the resource_creation_form, and instead of using the `view/components/form`
files, uses the forme library for generating the HTML.  Customization of
the forms for Ubicloud's needs is done in a single file,
`helpers/clover_forme_config.rb`.

Currently, this doesn't use a model-based form.  Using a model-based
form would likely make things even simpler, and could be done later, but
requires additional refactoring.