Switch firewall creation form to be a model form
Forme supports forms tied to Sequel::Model instances, simplifying form creation.
With model forms, forme can automatically determine an appropriate form input
name, id, label, whether the input is required, what the current value of
the input should be, if the input has any associated errors, etc.

The name, description, and location inputs in the firewall creation form can be
model inputs.  The private_subnet_id input cannot be, as that is not tied to
a column or association (there is a many_to_many association with private
subnets, but the form only allows the selection of a single subnet).

When using model forms, you have to have a model instance, so one is
instantiated before rendering the form.  This uses the Sequel and Roda
forme_set plugins, simplifying form intake.  When using the plugin,
Roda can determine from the submission which model input fields were
present when the form was created, and can automatically handle updating
the model instance.

This moves all validation for firewall creation to the model. A
Validation::ModelNameValidation module has been added for the name validation,
and Firewall now includes it.

Instead of having the error handler rescue validation failures, and preform
an internal request to the user-provided referrer (a horrifying practice),
this adds a handle_validation_failure helper method.  If there is a validation
failure inside the passed block, this sets a flash error for the current page,
then renders the given template.  Since all validation is in the model, and
forme knows how to handle model errors, things work without special code.

While here, enforce the requested private subnet is related to the current
project, and is in the same location as the firewall being created.