Convert committee pattern failures to HTTP 400 errors
The recent `openapi.yml` changes introduce
`components/schemas/Reference`, which validates UBIDs and names for a
resource. Malformed strings now raise exceptions that can be converted
to error content and codes as needed.

This change highlighted an issue: previously, DELETE always returned
204. This made sense for DELETE requests against a valid name or UBID,
but it was unclear what should happen when the string in the "name"
segment could never result in any action. It remained a 204, but
considering that other validation errors return 400, it makes more
sense for this case to also return 400—the resource can never exist,
and the DELETE cannot have ever succeeded.

Pleasingly, this slightly reduces code in `clover.rb`, where the
exception-to-HTTP conversion now aligns with other cases that yield
400, rather than being a special case emitting 204 on DELETE for no
clear reason.

Daniel Farina <daniel@ubicloud.com> wrote this commit message.