ConvergePostgresResource deadline change for maintenance windows
With the introduction of maintenance windows, the refresh of a PG
instance may simply wait for a while until the maintenance window comes.
At the start of the ConvergePostgresResource, we are setting the
deadline to 2 hours. This simply triggers a page if the preparation of a
server was quick but still we are waiting because we are not yet at the
maintenance window. With this commit, we are fixing that.