Persist invoices as PDFs in blob storage
Invoices are legal documents and should not be changed. Currently, we
generate the PDFs in real-time using the content of the invoices. We try
to keep it backward compatible and not change the format of the PDFs,
but that might happen over time. It's important to save the version of
the invoice that was sent to the customer.

I will also show the PDF invoice to the customer from blob storage, but
I will do that in a future PR after persisting all existing invoices.

We also need to share our generated invoices with our accountants for
VAT purposes. To make sharing easier, I place them in separate
directories.