Show persisted invoices from blob storage
We started persisting invoices as PDFs in blob storage. Since they are
legal documents, we should always show the same PDF to the customer.

Instead of generating the PDF on the fly, we can display it from the
saved invoices.

If the invoice doesn't exist in the blob storage, we will still generate
it at runtime to fulfill the customer's request.