Put PostgreSQL standbys to different data centers
We don't have different availability zones in none of our bare metal providers,
but at least, in Hetzner, we have different data centers. This commit ensures
that we put PostgreSQL standbys in different data centers if they are using
Hetzner as underlying provider.

For LeaseWeb, we don't even have different data centers, but we can place the
servers to different racks. Currently all our hosts are already in different
racks, so we don't need to do anything special for them.